% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylog.R
\name{mylog}
\alias{mylog}
\title{Compute log vector}
\usage{
mylog(vect)
}
\arguments{
\item{vect}{a vector of length n}
}
\value{
returns the vector zith log compted for each element
}
\description{
Compute log vector
}
\examples{
v <- c(1, 2, 3)
mylog(v)

}
